added discussion on POPL'19 paper
resolves #13

--- response letter

Thank you for pointing out the paper.  We admit that we were not aware
of the paper at the time of submission.  After reading the paper, we
concluded that the paper and our work are fundamentally different in the
domains of embedding and the applications.  The paper proposes a method
of representing one AST, whereas our work proposes a method representing
one token in an AST.  Nevertheless, there is interesting similarities
between the paper and our work with respect to "context" of tokens.  We
added a related work section that discuss those points.